# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TGS Admin API Documentation",
        "description": "Documentation de l'API REST pour le système TGS Admin - Toulouse Game Show",
        "contact": {
            "email": "support@tgs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://votre-domaine.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/ticket-contents": {
            "get": {
                "tags": [
                    "Ticket Contents"
                ],
                "summary": "Liste de tous les contenus de billets",
                "description": "Récupère tous les contenus de billets avec pagination et recherche optionnelle",
                "operationId": "getAllTicketContents",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche dans le nom du contenu",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "accès"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contenus de billets récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-contents/{ticketContent}": {
            "get": {
                "tags": [
                    "Ticket Contents"
                ],
                "summary": "Détail d'un contenu de billet",
                "description": "Récupère les informations d'un contenu de billet spécifique",
                "operationId": "getTicketContent",
                "parameters": [
                    {
                        "name": "ticketContent",
                        "in": "path",
                        "description": "ID du contenu de billet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contenu de billet récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ticket-contents/search": {
            "get": {
                "tags": [
                    "Ticket Contents"
                ],
                "summary": "Recherche de contenus de billets",
                "description": "Recherche de contenus de billets pour autocomplete (limite de 20 résultats)",
                "operationId": "searchTicketContents",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vip"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}